requirements.txt
fastapi
uvicorn
langchain-openai
langchain-community
langchain-pinecone
pinecone
python-dotenv
